Separate developer tests from user tests
The commit separates the developer tests from the user tests for
the gaussian database. This allows for better sustainability.
